Support search, captcha. Fix admin authentication, pagination.

+++ src/controller/admin/user/DeleteController.java
@@ -15,13 +15,13 @@ import java.util.Map;
 public class DeleteController extends AdminBaseController {
     @Override
     protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
-        this.authentication(req, resp);
+        if (!this.authentication(req, resp)) return;
         Map<String, String> param = this.param(req);
         int id = Integer.parseInt(param.get("id"));
         int result = UserDAO.getInstance().where(Column.check("id", "=", id)).delete();
         if (result > 0) {
             this.message(req, resp, "删除成功");
-        }else {
+        } else {
             this.message(req, resp, "删除失败");
         }
         this.redirect(req, resp, "admin/user/index");
